requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
click==8.0.4
colorama==0.4.4
dataclasses==0.8
fastapi==0.66.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.13.0
importlib-metadata==4.8.3
promise==2.3
pydantic==1.9.1
Rx==1.6.1
six==1.16.0
starlette==0.14.2
typing-extensions==4.1.1
uvicorn==0.14.0
zipp==3.6.0
